refactor: walk:tree phase related plugins with newly designed WalkTreePlugin interface
